Move Settings to common
