+++ client/Commands/Command.ts
@@ -1,7 +1,7 @@
 import * as ko from "knockout";
 
 import _ = require("lodash");
-import { Settings } from "../Settings/Settings";
+import { Settings } from "../../common/Settings";
 import { Store } from "../Utility/Store";
 import { GetLegacyKeyBinding } from "./GetLegacyKeyBinding";
